pyproject.toml
[tool.poetry]
name = "productlist"
version = "0.1.0"
description = ""
authors = ["Wesley Borges <djwesleyborges@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
django-ninja = "^0.22.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
